feat: InsightsPanel 컴포넌트 추출 및 대시보드 AI 인사이트 패널 구현 (closes #137)
- app/components/dashboard/InsightsPanel.tsx 신규 컴포넌트 생성
  - 트렌드(positive/negative/neutral 배지), 제품 제안(priority 색상), 사각지대, 다음 분석 제안 렌더링
  - 분석 3개 미만 시 비활성 안내, 로딩 상태, 24시간 캐시 표시 포함
- dashboard/page.tsx에서 Section 6 인라인 코드를 InsightsPanel로 교체
- Insights/Trend/ProductSuggestion 인터페이스를 InsightsPanel로 이전
